Rename ModelRenderer2 to ModelRenderer
